[CodeView] Healthy paranoia around strings
Make sure strings don't get too big for a record, truncate them if
need-be.

llvm-svn: 273710

+++ llvm/include/llvm/DebugInfo/CodeView/ListRecordBuilder.h
@@ -50,7 +50,7 @@ private:
     return ContinuationOffsets.empty() ? 0 : ContinuationOffsets.back();
   }
   size_t getLastContinuationEnd() const { return Builder.size(); }
-  unsigned getLastContinuationSize() const {
+  size_t getLastContinuationSize() const {
     return getLastContinuationEnd() - getLastContinuationStart();
   }
